fixed PHPStan errors

+++ src/Dibi/Drivers/FirebirdResult.php
@@ -44,12 +44,12 @@ class FirebirdResult implements Dibi\ResultDriver
 			: @ibase_fetch_row($this->resultSet, IBASE_TEXT); // intentionally @
 
 		if (ibase_errcode()) {
-			if (ibase_errcode() === FirebirdDriver::ERROR_EXCEPTION_THROWN) {
+			if ((int) ibase_errcode() === FirebirdDriver::ERROR_EXCEPTION_THROWN) {
 				preg_match('/exception (\d+) (\w+) (.*)/is', ibase_errmsg(), $match);
-				throw new Dibi\ProcedureException($match[3], $match[1], $match[2]);
+				throw new Dibi\ProcedureException($match[3], (int) $match[1], $match[2]);
 
 			} else {
-				throw new Dibi\DriverException(ibase_errmsg(), ibase_errcode());
+				throw new Dibi\DriverException(ibase_errmsg(), (int) ibase_errcode());
 			}
 		}
